+++ Polyclinic/Polyclinic.Domain/BloodGroup.cs
@@ -1,0 +1,12 @@
+﻿namespace Domain;
+
+/// <summary>
+/// Enumeration of patient blood group
+/// </summary>
+public enum BloodGroup
+{
+    I,
+    II,
+    III,
+    IV
+}
