+++ Polyclinic/Polyclinic.Domain/Doctor.cs
@@ -1,22 +1,6 @@
 ﻿namespace Domain;
 
 /// <summary>
-/// Enumeration of doctor specializations
-/// </summary>
-public enum Specialization
-{
-    Therapist,
-    Surgeon,
-    Cardiologist,
-    Neurologist,
-    Pediatrician,
-    Dentist,
-    Ophthalmologist,
-    Dermatologist,
-    Orthopedist,
-}
-
-/// <summary>
 /// Class describing a doctor
 /// </summary>
 public class Doctor
