+++ Polyclinic/Polyclinic.Domain/Gender.cs
@@ -1,0 +1,10 @@
+﻿namespace Domain;
+
+/// <summary>
+/// Enumeration of patient gender
+/// </summary>
+public enum Gender
+{
+    Male,
+    Female
+}
