+++ Polyclinic/Polyclinic.Domain/RhFactor.cs
@@ -1,0 +1,10 @@
+﻿namespace Domain;
+
+/// <summary>
+/// Enumeration of patient Rh factor
+/// </summary>
+public enum RhFactor
+{
+    Positive,
+    Negative
+}
