+++ Polyclinic/Polyclinic.Domain/Specialization.cs
@@ -1,0 +1,17 @@
+namespace Domain;
+
+/// <summary>
+/// Enumeration of doctor specializations
+/// </summary>
+public enum Specialization
+{
+    Therapist,
+    Surgeon,
+    Cardiologist,
+    Neurologist,
+    Pediatrician,
+    Dentist,
+    Ophthalmologist,
+    Dermatologist,
+    Orthopedist,
+}
